move CreateImageIndex parameters to one class
